Creating building js class.
