--- .coveralls.yml
@@ -1,1 +1,0 @@
-repo_token: togFVZcnCwIEezRsMHecWBzSIhr7dFS8x
