+++ darcyWeb/package.json
@@ -1,9 +1,11 @@
 {
   "dependencies": {
     "@rails/webpacker": "^3.0.2",
+    "axios": "^0.17.0",
     "coffeescript": "1.12.7"
   },
   "devDependencies": {
+    "babel-eslint": "^8.0.1",
     "eslint": "^4.10.0",
     "eslint-config-airbnb": "^16.1.0",
     "eslint-formatter-pretty": "^1.3.0",
