+++ docker-compose.yml
@@ -5,7 +5,7 @@ services:
     image: postgres
   webpacker:
     build: .
-    command: bash -c "rm -f tmp/pids/server.pid && bundle exec rails s -p 3000 -b '0.0.0.0'"
+    command: bundle exec bin/webpack-dev-server --host 0.0.0.0
     volumes:
       - .:/ondeeunb
     ports:
@@ -13,7 +13,7 @@ services:
 
   web:
     build: .
-    command: bundle exec rails s -p 3000 -b '0.0.0.0'
+    command: bash -c "rm -f tmp/pids/server.pid && bundle exec rails s -p 3000 -b '0.0.0.0'"
     volumes:
       - .:/ondeeunb
     ports:
